+++ src/builder/create_helper.cpp
@@ -2,7 +2,7 @@
 #include <NJS/Type.hpp>
 #include <NJS/Value.hpp>
 
-llvm::Value* NJS::Builder::CreateAlloca(llvm::Type* type, const size_t size) const
+llvm::Value* NJS::Builder::CreateAlloca(llvm::Type* type, const unsigned size) const
 {
     const auto bkp = GetBuilder().GetInsertBlock();
     GetBuilder().SetInsertPointPastAllocas(bkp->getParent());
@@ -11,7 +11,7 @@ llvm::Value* NJS::Builder::CreateAlloca(llvm::Type* type, const size_t size) con
     return ptr;
 }
 
-NJS::ValuePtr NJS::Builder::CreateAlloca(const TypePtr& type, const size_t size)
+NJS::ValuePtr NJS::Builder::CreateAlloca(const TypePtr& type, const unsigned size)
 {
     const auto ptr = CreateAlloca(type->GetLLVM(*this), size);
     return LValue::Create(*this, type, ptr);
